+++ NEWS.md
@@ -1,5 +1,18 @@
 Future TODO:
 
+v0.1.2 (Release date: 2026-09-08)
+=============================================
+- Added functions for visualization of clustering: 
+ -> showClusterTable: Present table
+ -> plotCommonCluster: Present igraph
+
+
+v0.1.1 (Release date: 2026-09-08)
+=============================================
+- Fixed genotype sorting issue in calcCommonCluster 
+ -> Added sorting at L80 and L156-L160 block
+
+
 v0.1.0 (Release date: 2026-09-07)
 =============================================
 Init release